添加主内存自动gc，添加过期对象，添加命令gc，添加过期对象回调方法

+++ src/main/java/cainsgl/redis/core/storage/RedisObjFactory.java
@@ -44,7 +44,7 @@ public class RedisObjFactory
     {
         try
         {
-            float v = Integer.parseInt(obj);
+            int v = Integer.parseInt(obj);
             return new RedisNumberObj(v);
         } catch (Exception e)
         {
